Let warmup be a fraction of an epoch
Warmup is converted to steps as warmup_epochs * steps_per_epoch, but the
flag was type=int, so the smallest nonzero value was a whole epoch. With
training_epochs 1 that made warmup consume the entire update: the
learning rate ramped from zero to peak across every step and the decay
phase never ran. The comment above the conversion already warned about
this shape; the type made it unavoidable.

ddrnet-deep and ddrnet-resign both ran that way. It matters more the
shorter the update: at 474 steps the ramp still reaches useful rates for
most of the pass, at 92 it does not.

+++ training/vgo_training/learner.py
@@ -95,7 +95,11 @@ class LearnerConfig:
     compile: bool = True
     restore_optimizer: bool = True
     schedule: str = "wsd"
-    warmup_epochs: int = 5
+    # Fractional values are meaningful and sometimes necessary: warmup is
+    # converted to steps as warmup_epochs * steps_per_epoch, so with a short
+    # epoch count an integer floor of 1 makes warmup consume the whole update
+    # and the decay phase never runs.
+    warmup_epochs: float = 5
     decay_fraction: float = 0.2
     final_learning_rate_fraction: float = 0.01
     report_every: int = 20
